pyproject.toml
# --------------------------------------------------------------------------- #
#                                BUILD SYSTEM                                 #
# --------------------------------------------------------------------------- #

[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools >=64.0.0",
    "wheel >=0.40.0",
]


# --------------------------------------------------------------------------- #
#                           PROJECT AND DEPENDENCIES                          #
# --------------------------------------------------------------------------- #

[project]
name = "diffusion-cli"
version = "0.0.1"
description = "diffusion_cli: Simple command line and graphical user interface for diffusion models."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
authors = [{name = "Jan Schlegel", email = "jschlegel@student.ethz.ch"}]
keywords = ["diffusion models", "machine learning"]
dependencies = [
    "torch>=1.9.0",
    "diffusers>=0.14.0",
    "transformers>=4.25.1",
    "accelerate>=0.16.0",
    "PyQt6>=6.4.0",
    "pyyaml>=6.0",
    "pillow>=9.3.0",
]

[project.optional-dependencies]
dev = [
    "black[d]",
    "coverage",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
    "ruff",
    "types-PyYAML",
    "pre-commit",
]


[project.scripts]
diffusion-cli = "diffusion_cli.cli:main"
diffusion-gui = "diffusion_cli.gui:run_gui"

[project.urls]
Homepage = "https://github.com/jhschlegel/diffusion-cli"
Issues = "https://github.com/jhschlegel/diffusion-cli/issues"


# --------------------------------------------------------------------------- #
#                          TOOL CONFIGURATION                                 #
# --------------------------------------------------------------------------- #

[tool.black]
line-length = 79
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_redundant_casts = true
warn_unreachable = false
ignore_missing_imports = true
disallow_untyped_calls = true
no_implicit_optional = true
disallow_untyped_defs = true

[tool.pyright]
pythonVersion = "3.10"
reportPrivateImportUsage = "none"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
testpaths = [
    "tests",
]
markers = [  # deselect tests with -m "not <marker>"
    "slow: unit tests that are slow",
]

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle rules
    "F",  # Pyflakes rules
    "B",  # flake8-bugbear rules
]


[tool.setuptools]
packages = ["diffusion_cli"]
package-dir = {"" = "src"}
